Add option to fetch holidays from a URL
Meant for use with an API endpoint that returns a JSON array of
holidays, in string ISO 8601 format.

If this option is used, holiday dates are cached for 5 mins. That way
changes to the data being returned by the calendar API will not
necessitate restarting every app that uses business\_calendar with that
API.